new patch for more stubs and misc changes patch changes: 1.add mswsock.dll with stubs for allmost all functions (1 or 2 I cant get prototype of) 2.add some definitions to various header files 3.add some stubs to advapi32.dll 4.add a stub for GdiInitializeLanguagePack to gdi32.dll 5.add some stubs to iphlpapi.dll 6.add some stubs to kernel32.dll 7.add some stubs to ntdll.dll 8.add some stubs to ws2_32.dll 9.add some stub definitions to wsock32.dll 10.change wsock32.dll so that most functions redirect to ws2_32.dll or mswsock.dll (which is how microsoft does it) 11.turn -Wall -Werror on for cards 12.turn -Wall -Werror on for epsapi 13.turn -Wall -Werror on for fmifs and fix problems that resulted 14.turn -Wall -Werror on for freetype 15.turn -Wall -Werror on for vfatlib and fix problems that resulted 15.turn -Wall -Werror on for kernel32 and fix problems that resulted 16.turn -Wall -Werror on for ntdll and fix problems that resulted 17.turn -Wall -Werror on for lzexpand 18.turn -Wall -Werror on for msafd 19.turn -Wall -Werror on for psapi 20.turn -Wall -Werror on for twain 21.add some stubs to user32.dll 22.turn -Wall -Werror on for version 23.turn -Wall -Werror on for wshirda
svn path=/trunk/; revision=6055

+++ reactos/lib/ntdll/rtl/nls.c
@@ -1,4 +1,4 @@
-/* $Id: nls.c,v 1.14 2003/07/26 23:25:17 royce Exp $
+/* $Id: nls.c,v 1.15 2003/09/12 17:51:48 vizzini Exp $
  *
  * COPYRIGHT:       See COPYING in the top level directory
  * PROJECT:         ReactOS kernel
@@ -309,7 +309,7 @@ RtlOemToUnicodeN (PWCHAR UnicodeString,
 
       for (i = 0; i < Size; i++)
 	{
-	  *UnicodeString = NlsOemToUnicodeTable[*OemString];
+	  *UnicodeString = NlsOemToUnicodeTable[(INT)*OemString];
 	  UnicodeString++;
 	  OemString++;
 	}
@@ -681,7 +681,7 @@ RtlUpperChar (IN CHAR Source)
       /* single-byte code page */
 
       /* ansi->unicode */
-      Unicode = NlsAnsiToUnicodeTable[Source];
+      Unicode = NlsAnsiToUnicodeTable[(INT)Source];
 
       /* upcase conversion */
       Unicode = RtlUpcaseUnicodeChar (Unicode);
